Silence clang-tidy warnings

* Silence various clang-tidy warnings

* More clang-tidy warnings, fix uninitialized members.

* Fix QuadraticCostFunction

* Correcting doc

Approved-by: Farbod Farshidian

+++ ocs2_comm_interfaces/src/ocs2_ros_interfaces/mpc/MPC_ROS_Interface.cpp
@@ -366,7 +366,7 @@ void MPC_ROS_Interface::spin() {
   } catch (...) {
     // declaring that MPC is not updated anymore
     ocs2_msgs::mpc_flattened_controller mpcPolicyMsg;
-    mpcPolicyMsg.controllerIsUpdated = false;
+    mpcPolicyMsg.controllerIsUpdated = static_cast<uint8_t>(false);
     mpcPolicyPublisher_.publish(mpcPolicyMsg);
     throw;
   }
